+++ lizard_riool/urls.py
@@ -36,6 +36,10 @@ urlpatterns = patterns(
     url(r'^archief/(?P<page_number>\d+)/$',
         login_required(views.ArchivePage.as_view()),
         name='lizard_riool_archive_page_numbered'),
+    url(r'^stelsels/(?P<sewerage_id>\d+)/generated_rib/$',
+        login_required(views.DownloadView.as_view()),
+        name='lizard_riool_download_generated_rib'),
+
     # Activate / deactivate and also deletion. One uses POST the other DELETE
     url(r'^stelsels/(?P<sewerage_id>\d+)/$',
         login_required(views.activate_sewerage_view),
@@ -49,8 +53,6 @@ urlpatterns = patterns(
             views.FileView.as_view(template_name="lizard_riool/files.html"))),
     url(r'^beheer/files/upload/$', login_required(views.UploadView.as_view()),
         name="upload_dialog_url"),
-    (r'^beheer/files/download/(?P<id>\d+)/$', login_required(
-            views.DownloadView.as_view())),
     (r'^beheer/files/delete/(?P<id>\d+)/$', login_required(
             views.DeleteFileView.as_view())),
     (r'^langsprofielen/$', login_required(views.SideProfileView.as_view())),